requirements.txt
attrs==25.3.0
capmonster_python==3.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
mysql-connector-python==9.0.0
nopecha==2.0.1
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.12.2
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
